move free related classes to more vertical package
